device and swapchain.

+++ Engine/src/Window.h
@@ -21,8 +21,8 @@ public:
     // Use when initializing the window.
     struct Desc
     {
-        int width = GetSystemMetrics(SM_CXSCREEN) / 2;
-        int height = GetSystemMetrics(SM_CYSCREEN) / 2;
+        unsigned int width = (unsigned int)GetSystemMetrics(SM_CXSCREEN) / 2;
+        unsigned int height = (unsigned int)GetSystemMetrics(SM_CYSCREEN) / 2;
         LPCWSTR title = L"Default";
         bool fullScreen = false;
         bool resizableWindow = false;
@@ -36,8 +36,8 @@ public:
 	
     HWND getHWnd() const;
     RECT getClientRect() const;
-    INT getWidth() const;
-    INT getHeight() const;
+    unsigned int getWidth() const;
+    unsigned int getHeight() const;
 
 private:	
     HWND hWnd;
